Update to use new library package
