+++ package.json
@@ -7,19 +7,16 @@
     "build": "npm run build-ts",
     "build-ts": "tsc",
     "lint": "eslint . --ext .js,.ts",
-    "serve": "node dist/server.js",
+    "serve": "npm run build && node dist/server.js",
     "start": "npm run serve",
     "test": "ts-node node_modules/jasmine/bin/jasmine"
   },
   "author": "",
   "license": "Apache-2.0",
   "dependencies": {
-    "@types/xml2json": "^0.11.0",
     "body-parser": "^1.19.0",
+    "clinical-trial-matching-service": "github:mcode/clinical-trial-matching-service#initial-implementation",
     "express": "^4.17.1",
-    "fs": "0.0.1-security",
-    "jasmine": "^3.5.0",
-    "node-fetch": "^2.6.0",
     "xml2json": "^0.12.0"
   },
   "devDependencies": {
@@ -27,11 +24,13 @@
     "@types/jasmine": "^3.5.10",
     "@types/node": "^14.0.13",
     "@types/supertest": "^2.0.9",
+    "@types/xml2json": "^0.11.0",
     "@typescript-eslint/eslint-plugin": "^3.4.0",
     "@typescript-eslint/parser": "^3.4.0",
     "eslint": "^6.8.0",
     "eslint-config-prettier": "^6.11.0",
     "eslint-plugin-prettier": "^3.1.4",
+    "jasmine": "^3.5.0",
     "prettier": "2.0.5",
     "supertest": "^4.0.2",
     "ts-node": "^8.10.2",
